rust: move 'rust_common_flags' to a separate configuration file
Currently, 'rustc_common_flags' lives in the top-level 'Makefile'.
This means that we need to touch this whenever adjustments to Rust's CLI
configuration are desired, such as changing what lints Clippy should
emit.

This patch moves these flags to a separate file within 'rust/' so that
necessary changes can be made without affecting 'Makefile'. It copies
the behavior currently used with 'bindgen_parameters', which accepts
comments.

Additionally, 'let_unit_value` is no longer specifically named as it is
now part of the 'style' group [1].

[1]: https://github.com/rust-lang/rust-clippy/pull/8563

Signed-off-by: Trevor Gross <tmgross@umich.edu>